Add freetext list for ingredients

+++ Cookbook/ShoppingListView.swift
@@ -13,14 +13,58 @@ struct ShoppingListView: View {
     
     @Query() var shoppingLists: [ShoppingList]
     
+    @State var addShowing: Bool = false
+    
+    @FocusState var keyboardIsActive: Bool
+    @StateObject var ingredient: Ingredient = Ingredient(name: "", quantityWhole: 1, quantityFraction: 0, unit: "item")
+    
     var body: some View {
         NavigationStack {
             @Bindable var shoppingList = shoppingLists.first!
-            List {
-                IngredientListSection(ingredients: shoppingList.getItems(), selections: $shoppingList.selections, onDelete: { indexSet in
-                    shoppingList.deleteItem(indexSet: indexSet, context: context)
-                })
+            
+            Form {
+                
+                if shoppingList.getItems().count > 0 {
+                    IngredientListSection(ingredients: shoppingList.getItems(), selections: $shoppingList.selections, onDelete: { indexSet in
+                        shoppingList.deleteItem(indexSet: indexSet, context: context)
+                    })
+                }
+                
+                ListButton(text: "Add Ingredients", imageSystemName: "plus") {
+                    addShowing = true
+                }
             }
+            .onScrollPhaseChange({ oldPhase, newPhase in
+                if (oldPhase == .interacting) {
+                    keyboardIsActive = false
+                }
+            })
+            .sheet(isPresented: $addShowing, content: {
+                NavigationStack {
+                    Form {
+                        CreateEditIngredient(ingredient: ingredient, onSubmit: {
+                            shoppingList.addItem(ingredient)
+                            ingredient.name = ""
+                            ingredient.quantityWhole = 1
+                            ingredient.quantityFraction = 0
+                            ingredient.unit = "item"
+                            keyboardIsActive = true
+                        }, keyboardIsActive: $keyboardIsActive)
+                        
+                        IngredientListSection(ingredients: shoppingList.getItems(), selections: $shoppingList.selections, onDelete: { indexSet in
+                            shoppingList.deleteItem(indexSet: indexSet, context: context)
+                        })
+                    }
+                    .toolbar {
+                        Button {
+                            addShowing = false
+                        } label: {
+                            Text("Done")
+                        }
+                    }
+                }
+                
+            })
             .navigationTitle("Shopping List")
             .toolbar {
                 Button {
@@ -30,6 +74,5 @@ struct ShoppingListView: View {
                 }
             }
         }
-        
     }
 }
